[Rector] rectorfy code

+++ src/CoreShop/Component/Resource/TokenGenerator/UniqueTokenGenerator.php
@@ -61,7 +61,7 @@ final class UniqueTokenGenerator
         $log = log($range, 2);
         $bytes = (int) ($log / 8) + 1;
         $bits = (int) $log + 1;
-        $filter = (int) (1 << $bits) - 1;
+        $filter = (1 << $bits) - 1;
 
         do {
             $rnd = hexdec(bin2hex(openssl_random_pseudo_bytes($bytes)));
